+++ CHANGELOG.md
@@ -7,7 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased](https://github.com/orisai/nette-auth/compare/1.0.1...HEAD)
 
+### Fixed
+
+- SessionLoginStorage
+	- supports namespace with numeric name
+
 ## [1.0.1](https://github.com/orisai/nette-auth/compare/1.0.0...1.0.1) - 2022-12-09
+
+### Changed
 
 - Composer
 	- allows PHP 8.2